pico-mp: write multiple tests to figure fader control
Tinker with the faders, with the goal of having great control of moving
them wherever.

Write some test code to make it follow a separate potentiometer.

Write some test code to make it go to hardcoded positions.

Implement a proportional (P) motor controller, as a start for
implementing a PID controller (probably unnecessary here, but will be a
good learning exercise).

Currently, the ADC inputs have to be shifted down to 6 bits. This is
really bad and wasn't an issue when I tried something like this years
ago on an Arduino Uno. It seems the Pico ADC is way too noise, so will
need to try using a shunt reference or potentially an external ADC later
on.

For now, this is good enough to continue figuring out the next stages.
The next logical step would probably be to figure out how to control
faders individually through a multiplexer and/or shift register.