pico-cp: link Kronos fader with the motorised one
Move ported motor-movement code to a separate file for future reference.

Slightly modify the `KronosMessage` parser to stop parsing if a MIDI
packet is empty (prevent it getting stuck in the loop when it receives
any other data).

Receive and parse Kronos fader-movement data, and make the motorised
fader move to that position.

The data is received/parsed (not sure which one of the two yet) with a
significant delay and MIDI packets get dropped. This will have to be
looked into. Not sure if CircuitPython is fast enough, or if there is a
major inefficiency in the code that causes this.

Being a derisk, I call this a win!